Move plugin_common.[ch] to a common subdir.
Doing this will fix issues with distclean.
