+++ common/plugin_common.c
