+++ common/plugin_common.h
