+++ configure.ac
@@ -1582,6 +1582,7 @@ libsasl2.pc
 include/Makefile
 saslauthd/Makefile
 sasldb/Makefile
+common/Makefile
 plugins/Makefile
 lib/Makefile
 utils/Makefile
