+++ lib/NTMakefile
@@ -52,11 +52,11 @@ libsasl.dll: $(libsasl_objs) $(libsasl_res)
 <<
 	IF EXIST $@.manifest mt -manifest $@.manifest -outputresource:$@;2
 
-plugin_common.c: ..\plugins\plugin_common.c plugin_common.h
-	xcopy /D /Y ..\plugins\plugin_common.c .
+plugin_common.c: ..\common\plugin_common.c plugin_common.h
+	xcopy /D /Y ..\common\plugin_common.c .
 
-plugin_common.h: ..\plugins\plugin_common.h
-	xcopy /D /Y ..\plugins\plugin_common.h .
+plugin_common.h: ..\common\plugin_common.h
+	xcopy /D /Y ..\common\plugin_common.h .
 
 auxprop.obj checkpw.obj client.obj common.obj external.obj plugin_common.obj server.obj seterror.obj: ..\include\saslplug.h
